tidy up utils about types
