+++ canals/pipeline/draw/draw.py
@@ -11,7 +11,7 @@ import networkx
 from canals.pipeline.validation import _find_pipeline_inputs, _find_pipeline_outputs
 from canals.pipeline.draw.graphviz import _to_agraph
 from canals.pipeline.draw.mermaid import _to_mermaid_image, _to_mermaid_text
-from canals.utils import _type_name
+from canals.type_utils import _type_name
 
 logger = logging.getLogger(__name__)
 RenderingEngines = Literal["graphviz", "mermaid-image", "mermaid-text"]
