+++ canals/type_utils.py
@@ -9,6 +9,13 @@ import logging
 logger = logging.getLogger(__name__)
 
 
+def _is_optional(type_: type) -> bool:
+    """
+    Utility method that returns whether a type is Optional.
+    """
+    return get_origin(type_) is Union and type(None) in get_args(type_)
+
+
 def _types_are_compatible(sender, receiver):  # pylint: disable=too-many-return-statements
     """
     Checks whether the source type is equal or a subtype of the destination type. Used to validate pipeline connections.
@@ -46,3 +53,31 @@ def _types_are_compatible(sender, receiver):  # pylint: disable=too-many-return-
         return False
 
     return all(_types_are_compatible(*args) for args in zip(sender_args, receiver_args))
+
+
+def _type_name(type_):
+    """
+    Util methods to get a nice readable representation of a type.
+    Handles Optional and Literal in a special way to make it more readable.
+    """
+    # Literal args are strings, so we wrap them in quotes to make it clear
+    if isinstance(type_, str):
+        return f"'{type_}'"
+
+    name = getattr(type_, "__name__", str(type_))
+
+    if name.startswith("typing."):
+        name = name[7:]
+    if "[" in name:
+        name = name.split("[")[0]
+    args = get_args(type_)
+    if name == "Union" and type(None) in args and len(args) == 2:
+        # Optional is technically a Union of type and None
+        # but we want to display it as Optional
+        name = "Optional"
+
+    if args:
+        args = ", ".join([_type_name(a) for a in args if a is not type(None)])
+        return f"{name}[{args}]"
+
+    return f"{name}"
